added custom actions and modified generated views

+++ Controllers/InventoryController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using BooksOnLoan.Data;
 using BooksOnLoan.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace BooksOnLoan.Controllers
 {
@@ -19,21 +20,55 @@ namespace BooksOnLoan.Controllers
             _context = context;
         }
 
+        [Authorize]
         // GET: Inventory
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Inventory.ToListAsync());
+            var query = _context.Transactions!
+                .Where(t => t.ReturnDate == null)
+                .GroupBy(t => t.BookCodeNumber)                
+                .Select(g => new
+                {
+                    Book = g.Key,
+                    TransactionCount = g.Count()
+                });
+
+            int overdueCount = _context.Transactions!
+                .Where(t => t.ReturnDate == null)
+                .Where(t => t.LoanDueDate < DateOnly.FromDateTime(DateTime.Now))
+                .Where(t => t.Username == User.Identity!.Name)
+                .Count();
+
+            ViewBag.OverdueCount = overdueCount;
+
+            Dictionary<int, int> bookTransactionCounts = new Dictionary<int, int>();
+            foreach (var item in query)
+            {
+                bookTransactionCounts.Add(item.Book, item.TransactionCount);
+            }
+
+            ViewBag.BookTransactionCounts = bookTransactionCounts;
+
+            return View(await _context.Inventory!.ToListAsync());
         }
 
+        [Authorize]
         // GET: Inventory/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (User.IsInRole("Admin") is false)
+            {
+                return NotFound();
+            }
+
+
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var book = await _context.Inventory
+            var book = await _context.Inventory!
                 .FirstOrDefaultAsync(m => m.CodeNumber == id);
             if (book == null)
             {
@@ -43,19 +78,35 @@ namespace BooksOnLoan.Controllers
             return View(book);
         }
 
+        [Authorize]
         // GET: Inventory/Create
         public IActionResult Create()
         {
+            if (User.IsInRole("Admin") is false)
+            {
+                return NotFound();
+            }
+
+
+
             return View();
         }
 
         // POST: Inventory/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("CodeNumber,Title,Author,YearPublished,Quantity")] Book book)
         {
+            if (User.IsInRole("Admin") is false)
+            {
+                return NotFound();
+            }
+
+
+
             if (ModelState.IsValid)
             {
                 _context.Add(book);
@@ -66,14 +117,22 @@ namespace BooksOnLoan.Controllers
         }
 
         // GET: Inventory/Edit/5
+        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
+            if (User.IsInRole("Admin") is false)
+            {
+                return NotFound();
+            }
+
+
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var book = await _context.Inventory.FindAsync(id);
+            var book = await _context.Inventory!.FindAsync(id);
             if (book == null)
             {
                 return NotFound();
@@ -84,10 +143,18 @@ namespace BooksOnLoan.Controllers
         // POST: Inventory/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("CodeNumber,Title,Author,YearPublished,Quantity")] Book book)
         {
+            if (User.IsInRole("Admin") is false)
+            {
+                return NotFound();
+            }
+
+
+
             if (id != book.CodeNumber)
             {
                 return NotFound();
@@ -117,14 +184,22 @@ namespace BooksOnLoan.Controllers
         }
 
         // GET: Inventory/Delete/5
+        [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
+            if (User.IsInRole("Admin") is false)
+            {
+                return NotFound();
+            }
+
+
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var book = await _context.Inventory
+            var book = await _context.Inventory!
                 .FirstOrDefaultAsync(m => m.CodeNumber == id);
             if (book == null)
             {
@@ -137,9 +212,17 @@ namespace BooksOnLoan.Controllers
         // POST: Inventory/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var book = await _context.Inventory.FindAsync(id);
+            if (User.IsInRole("Admin") is false)
+            {
+                return NotFound();
+            }
+
+
+
+            var book = await _context.Inventory!.FindAsync(id);
             if (book != null)
             {
                 _context.Inventory.Remove(book);
@@ -151,7 +234,7 @@ namespace BooksOnLoan.Controllers
 
         private bool BookExists(int id)
         {
-            return _context.Inventory.Any(e => e.CodeNumber == id);
+            return _context.Inventory!.Any(e => e.CodeNumber == id);
         }
     }
 }
